+++ slash/plugins/builtin/notifications.py
@@ -33,6 +33,9 @@ class Message(object):
 
 
 class Plugin(PluginInterface):
+    """Enables notifications for successful and failed test runs through various providers (NMA, Prowl, Pushover etc.)
+    """
+
     def get_name(self):
         return "notifications"